requirements-behave.txt
behave==1.2.6
browsermob-proxy==0.8.0
certifi==2018.4.16
chardet==3.0.4
chromedriver==2.24.1
chromedriver-installer==0.0.6
click==6.7
dominate==2.3.1
Flask==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Script==2.0.6
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
parse==1.9.0
parse-type==0.4.2
PyYAML==3.13
requests==2.18.4
selenium==3.14.0
six==1.11.0
urllib3==1.22
visitor==0.1.3
Werkzeug==0.12.2
xmlrunner==1.7.7
